[ADD] added details to map

+++ src/config_management/constructors/construct_map/map_parser.c
@@ -70,7 +70,7 @@ void display_map(sfRenderWindow *window, map_t *map_data)
     for (int i = 0; i < 34; i++) {
         for (int j = 0; j < 60; j++) {
             if (map_data->tiles[i][j] == 59)
-                map_data->tiles[i][j] = rdm_float(4, 8);
+                map_data->tiles[i][j] = (int)rdm_float(4, 8);
             limit.left = map_data->tiles[i][j] * TILE_SIZE;
             pos.x = (j * TILE_SIZE);
             pos.y = (i * TILE_SIZE);
